Remove changes to GetIssueLicenseURL from 87044 (bringing back into main card files) as the hooks won't work conditionally, as called in this lightbox
