+++ client/jetpack-cloud/sections/manage/pricing/license-products-list/simple-license-item-card.tsx
@@ -1,9 +1,10 @@
 import page from '@automattic/calypso-router';
 import { Button } from '@automattic/components';
 import { Icon, plugins } from '@wordpress/icons';
+import { addQueryArgs } from '@wordpress/url';
 import { useTranslate } from 'i18n-calypso';
 import { useCallback, useState } from 'react';
-import { useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
 import 'calypso/my-sites/plans/jetpack-plans/product-store/featured-item-card/style.scss';
 import getAPIFamilyProductIcon from 'calypso/jetpack-cloud/sections/manage/pricing/utils/get-api-family-product-icon';
 import { useProductDescription } from 'calypso/jetpack-cloud/sections/partner-portal/hooks';
@@ -12,9 +13,11 @@ import { LICENSE_INFO_MODAL_ID } from 'calypso/jetpack-cloud/sections/partner-po
 import getProductShortTitle from 'calypso/jetpack-cloud/sections/partner-portal/lib/get-product-short-title';
 import LicenseLightbox from 'calypso/jetpack-cloud/sections/partner-portal/license-lightbox/index';
 import LicenseLightboxLink from 'calypso/jetpack-cloud/sections/partner-portal/license-lightbox-link/index';
+import { useDispatch } from 'calypso/state';
 import { recordTracksEvent } from 'calypso/state/analytics/actions';
+import { isUserLoggedIn } from 'calypso/state/current-user/selectors';
+import { isAgencyUser } from 'calypso/state/partner-portal/partner/selectors';
 import { APIProductFamilyProduct } from 'calypso/state/partner-portal/types';
-import GetIssueLicenseURL from './get-issue-license-url';
 import { ItemPrice } from './item-price';
 
 type SimpleLicenseItemCardProps = {
@@ -37,6 +40,8 @@ export const SimpleLicenseItemCard = ( {
 }: SimpleLicenseItemCardProps ) => {
 	const translate = useTranslate();
 	const dispatch = useDispatch();
+	const isLoggedIn = useSelector( isUserLoggedIn );
+	const isAgency = useSelector( isAgencyUser );
 	const { setParams, resetParams, getParamValue } = useURLQueryParams();
 	const modalParamValue = getParamValue( LICENSE_INFO_MODAL_ID );
 	const [ showLightbox, setShowLightbox ] = useState( modalParamValue === item.slug );
@@ -101,9 +106,27 @@ export const SimpleLicenseItemCard = ( {
 		[ dispatch, item.slug, setParams ]
 	);
 
+	const getIssueLicenseURL = useCallback(
+		( productSlug: string, bundleSize: number | undefined ) => {
+			if ( isLoggedIn && ! isAgency ) {
+				return addQueryArgs( `/manage/signup/`, {
+					product_slug: productSlug,
+					source: 'manage-pricing-page',
+					bundle_size: bundleSize,
+				} );
+			}
+			return addQueryArgs( `/partner-portal/issue-license/`, {
+				product_slug: productSlug,
+				source: 'manage-pricing-page',
+				bundle_size: bundleSize,
+			} );
+		},
+		[ isLoggedIn, isAgency ]
+	);
+
 	const onSelectProduct = useCallback( () => {
-		page( GetIssueLicenseURL( productSlug, bundleSize ) );
-	}, [ bundleSize, productSlug ] );
+		page( getIssueLicenseURL( productSlug, bundleSize ) );
+	}, [ bundleSize, getIssueLicenseURL, productSlug ] );
 
 	const onHideLightbox = useCallback( () => {
 		resetParams( [ LICENSE_INFO_MODAL_ID ] );
@@ -128,7 +151,7 @@ export const SimpleLicenseItemCard = ( {
 							className="simple-item-card__cta"
 							onClick={ onClickCta }
 							disabled={ isCtaDisabled }
-							href={ isCtaDisabled ? '#' : GetIssueLicenseURL( productSlug, bundleSize ) }
+							href={ isCtaDisabled ? '#' : getIssueLicenseURL( productSlug, bundleSize ) }
 							target={ isCtaExternal ? '_blank' : undefined }
 							primary={ ctaAsPrimary }
 							aria-label={ ctaAriaLabel }
