+++ client/jetpack-cloud/sections/manage/pricing/license-products-list/simple-license-multi-item-card.tsx
@@ -1,8 +1,10 @@
 import { recordTracksEvent } from '@automattic/calypso-analytics';
 import page from '@automattic/calypso-router';
 import { Button } from '@automattic/components';
+import { addQueryArgs } from '@wordpress/url';
 import { useTranslate } from 'i18n-calypso';
 import { useCallback, useState } from 'react';
+import { useSelector } from 'react-redux';
 import MultipleChoiceQuestion from 'calypso/components/multiple-choice-question';
 import getAPIFamilyProductIcon from 'calypso/jetpack-cloud/sections/manage/pricing/utils/get-api-family-product-icon';
 import { useProductDescription } from 'calypso/jetpack-cloud/sections/partner-portal/hooks';
@@ -13,8 +15,9 @@ import getProductVariantShortTitle from 'calypso/jetpack-cloud/sections/partner-
 import LicenseLightbox from 'calypso/jetpack-cloud/sections/partner-portal/license-lightbox/index';
 import LicenseLightboxLink from 'calypso/jetpack-cloud/sections/partner-portal/license-lightbox-link/index';
 import { useDispatch } from 'calypso/state';
+import { isUserLoggedIn } from 'calypso/state/current-user/selectors';
+import { isAgencyUser } from 'calypso/state/partner-portal/partner/selectors';
 import { APIProductFamilyProduct } from 'calypso/state/partner-portal/types';
-import GetIssueLicenseURL from './get-issue-license-url';
 import { ItemPrice } from './item-price';
 import 'calypso/my-sites/plans/jetpack-plans/product-store/simple-item-card/style.scss';
 import './style.scss';
@@ -39,6 +42,8 @@ export const SimpleLicenseMultiItemCard = ( {
 }: SimpleLicenseMultiItemCardProps ) => {
 	const translate = useTranslate();
 	const dispatch = useDispatch();
+	const isLoggedIn = useSelector( isUserLoggedIn );
+	const isAgency = useSelector( isAgencyUser );
 	const { setParams, resetParams, getParamValue } = useURLQueryParams();
 	const modalParamValue = getParamValue( LICENSE_INFO_MODAL_ID );
 	const [ variant, setVariant ] = useState( variants[ 0 ] );
@@ -104,9 +109,27 @@ export const SimpleLicenseMultiItemCard = ( {
 		[ dispatch, variantSlug, setParams ]
 	);
 
+	const getIssueLicenseURL = useCallback(
+		( variantSlug: string, bundleSize: number | undefined ) => {
+			if ( isLoggedIn && ! isAgency ) {
+				return addQueryArgs( `/manage/signup/`, {
+					product_slug: variantSlug,
+					source: 'manage-pricing-page',
+					bundle_size: bundleSize,
+				} );
+			}
+			return addQueryArgs( `/partner-portal/issue-license/`, {
+				product_slug: variantSlug,
+				source: 'manage-pricing-page',
+				bundle_size: bundleSize,
+			} );
+		},
+		[ isLoggedIn, isAgency ]
+	);
+
 	const onSelectProduct = useCallback( () => {
-		page( GetIssueLicenseURL( variantSlug, bundleSize ) );
-	}, [ bundleSize, variantSlug ] );
+		page( getIssueLicenseURL( variantSlug, bundleSize ) );
+	}, [ bundleSize, getIssueLicenseURL, variantSlug ] );
 
 	const onHideLightbox = useCallback( () => {
 		resetParams( [ LICENSE_INFO_MODAL_ID ] );
@@ -129,7 +152,7 @@ export const SimpleLicenseMultiItemCard = ( {
 							className="simple-item-card__cta"
 							onClick={ onClickCta }
 							disabled={ isCtaDisabled }
-							href={ isCtaDisabled ? '#' : GetIssueLicenseURL( variantSlug, bundleSize ) }
+							href={ isCtaDisabled ? '#' : getIssueLicenseURL( variantSlug, bundleSize ) }
 							target={ isCtaExternal ? '_blank' : undefined }
 							primary={ ctaAsPrimary }
 							aria-label={ ctaAriaLabel }
